rename views -> ui
