+++ src/ui/daily.rs
@@ -7,7 +7,7 @@ use ratatui::{
 
 use anyhow::Result;
 
-use crate::views::common::{
+use crate::ui::common::{
     focusable::Focusable,
     view::{FocusableView, View},
 };
